Added authorization, fixed images uploading

+++ src/Components/BackEndApi.js
@@ -1,4 +1,5 @@
 export async function GetData(dbSetName, id = "", filter = "") {
+  if (id === null) return;
   const url =
     "http://localhost:5157/api/" +
     dbSetName +
@@ -12,39 +13,67 @@ export async function UpdateData(params) {
   const id = params.id;
   const body = JSON.stringify(params.body);
 
+  if (id === null) return;
+
   const url =
     "http://localhost:5157/api/" + dbSetName + (id !== "" ? "/" + id : "");
   return await fetch(url, {
     method: "PUT",
     headers: {
       "Content-Type": "application/json",
+      Authorization:
+        "Bearer " + JSON.parse(sessionStorage.getItem("AccessToken")),
     },
     body: body,
-  }).then((data) => data.json());
+  }).then((data) => data);
 }
 
 export async function AddData(params) {
   const dbSetName = params.dbSetName;
   const body = JSON.stringify(params.body);
-
   const url = "http://localhost:5157/api/" + dbSetName;
   return await fetch(url, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
+      Authorization:
+        "Bearer " + JSON.parse(sessionStorage.getItem("AccessToken")),
     },
     body: body,
   }).then((data) => data.json());
 }
 
+export async function AttachAppFile(params, bookId) {
+  if (bookId === null) return;
+  const dbSetName = params.dbSetName;
+  const body = params.body;
+  const url =
+    "http://localhost:5157/api/" + dbSetName + "/addApplicationFile/" + bookId;
+  return await fetch(url, {
+    method: "POST",
+    Authorization:
+      "Bearer " + JSON.parse(sessionStorage.getItem("AccessToken")),
+    body: body,
+  }).then((data) => {
+    data.json();
+  });
+}
+
 export async function DeleteData(params) {
   const dbSetName = params.dbSetName;
   const id = params.id;
+  if (id === null) return;
   const url =
     "http://localhost:5157/api/" + dbSetName + (id !== "" ? "/" + id : "");
+  console.log(
+    "DELETE Authorization >>> ",
+    "Bearer " + JSON.parse(sessionStorage.getItem("AccessToken"))
+  );
   return await fetch(url, {
     method: "DELETE",
-  }).then((data) => data.json());
+    Authorization:
+      "Bearer " + JSON.parse(sessionStorage.getItem("AccessToken")),
+  }).then((data) => data);
 }
 
 export async function Initialize(
